Add AnimalClickable & offset in click area
- Add AnimalClickable methods(Showing, Checking).
- Add offset into AnimalClickable and UIClickable for showing boundary and checking of clickable area.
- Modify UI.pde -> UserInterface.pde since the UI is not valid name for custom pde file(In my opinion, not exact).
- Modify position variables' type from int to float.

+++ RescueAnimal/Crab.pde
@@ -12,10 +12,10 @@ class Crab extends Animal {
   Crab() {
     super();
   }
-  Crab(int x, int y) {
+  Crab(float x, float y) {
     super(x, y);
   }
-  Crab(int x, int y, float rotation) {
+  Crab(float x, float y, float rotation) {
     super(x, y, rotation);
   }
